refactor(lib): refactor global skylark functions into modules
Things could get unruly rather quickly with the number of functions and modules
we're hoping to ship, so it makes sense that we start to work on a standard library.

I've taken an initial crack at setting such a library up, with two modules to get started:
'qri' and 'http'. Both of them now load with `load('qri.sky', 'qri')` and `load('http.sky', 'http')`