refactor(transform funcs): switch to passed-in environment dict
after some thought & conversation, we've decided to go a different route with transform functions. Instead of using load methods & protectors, we'll provide each transform func with an "environment object" that comes loaded with the things they can do within the context of that transform function. We think this'll make it easier to reason about what can happen when.