feat(html): added quick jquery-like html package
so, uh, this isn't exactly the best, but I'd really like to at least have a stand-in implementation for working with html documents in a sane way. I've never used beautifulsoup, and I'm guessing this'll aanger many pythonists, but to me jquery-style syntax / css selectors are the single best way to slice & dice html tags, so it's a start.