feat(OutWriter): accept writer param on ExecTransform to record script stdout
Providing an io.Writer value will now record output to the provided writer, allowing callers to capture a record of the script's output as it's being executed. Output is automatically piped to both Node.LocalStreams _and_ OutWriter if both are provided. A convenience method: `SetOutWriter` provides plug-and-play config for a given writer. It'll be a common pattern to allocate a *bytes.Buffer and pass that to SetOutWriter(buf), then read back the buffered output.