fix(ds): ds.set_body inherits structure instead of inventing one
set_body wasn't doing the right thing when it comes to determining the
destination datset structure for writing a body. We now inherit structure
from existing dataset data, and only fall back to json if no structure
exists.

The general problem here is a separation of concerns between data in the
starlark runtime and dataset serialization formats. Those should always
be seperate.

closes qri-io/qri#756